Purge 'dev' & 'test' secret keys
The secret key has *never* been revealed for 'production', but
there might still be worries about revealing secret keys for
'development' and 'test' database, despite the fact that these
databases are *never* supposed to be utilized in the public
website in the first place.

This change is made to address that worry. 

NOTE: Those previous 'dev' / 'test' keys are still in the 
repository's history! If you are using them, consider them
to be compromised and generate a new key with `rake secret`.